A new approach to the baseURI issue. Proposal: Add baseUri to the HttpServer config. * Base URI's can be transport specific. An http base URI might be different than the coap one, unless it is a proxy from one to the other. * A TD might be exposed over multiple transports, so the TD base attribute as it is, doesn't seem to cover that case * You may also have one HttpServer with multiple TD's exposed on it.  eg http://server.com/REDWOT and BLUEWOT So it feels like this belongs on the transport config rather than the TD, at least the way node-wot works.
I've also turned away from an ENV variable for the value.  At least in my experience, there tends to be a run script type of thingy
for rails and node apps that wrap the app code, take command line args, or mate with the deployment infrastructure like Heroku.  It's on the outer ring.  Putting it inside the http binding seems to hide it.  Like the PORT issue, it took a lot of time to troubleshoot because node-wot never told the user it overrode the setting.

Anyway... thoughts?

PS: this is based on master
Signed-off-by: Josh Cohen <joshco@gmail.com>